split off captureLogs, injectBundle, createDomReplayer from Page to helpers
