+++ packages/web-tests/tests/isSelectValueAvailable.test.ts
@@ -1,10 +1,10 @@
-import { Browser, Page } from "@qawolf/browser";
+import { Browser, PuppeteerPage } from "@qawolf/browser";
 import { CONFIG } from "@qawolf/config";
 import { QAWolfWeb } from "@qawolf/web";
 
 describe("isSelectValueAvailable", () => {
   let browser: Browser;
-  let page: Page;
+  let page: PuppeteerPage;
 
   beforeAll(async () => {
     browser = await Browser.create({ url: `${CONFIG.testUrl}dropdown` });
